refactor search & settings store tests #377
