+++ tests/profiles/store/search/actions/async.test.js
@@ -1,8 +1,8 @@
 import configureMockStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
 import nock from 'nock'
-import { SearchActions } from '../../../../app/js/profiles/store/search'
-import searchResults from '../../../fixtures/search'
+import { SearchActions } from '../../../../../app/js/profiles/store/search'
+import searchResults from '../../../../fixtures/search'
 
 const middlewares = [thunk]
 const mockStore = configureMockStore(middlewares)
